Javafied version of our C# library
Compared to the C# library, some functionality has been changed and some
left out. This due to the limitations in Java, particularly in the
inter-operability of generic structures and enums (or lack there of).
The generic structure now only serves as a superclass for the lists of
the individual enumerator types. Any methods that needs to access the
actual type of the "T" variable, needs to be moved to the subclass for
all enumerator list types.
If you think it sounds hard to grock, don't fret... it's just because it
is.

However... Don't worry, the usage is demonstrated in the test project.