WS - Match
getMatches now works... I think. Anyway, it isn't the most elegant
solution, but it will have to do for now. The first two solutions I
tried out (which worked perfectly and elegantly in the test
application). Wouldn't work in the WebService (limited framework
maybe?), so the implemeted solution is a bit brute-forcy and maybe also
a bit hard to grok.